+++ CHANGELOG.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-## v4.4.7 - 2025-03-26 - 4c548f037 -
+## v4.4.7 - 2025-03-26 - 4c548f037 - live since 2025-04-03
 
 - Fix #2241: Restore Sketchfab visualisations alongside 3D models files visualisations
 - Fix #2180: Add user in curation log when minting the DOI